pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fedps"
version = "0.1.0a0"
dependencies = [
    "scikit-learn>=1.3",
    "numpy>=1.20",
    "datasketches",
    "pyzmq",
]
requires-python = ">=3.8"
authors = [
    {name = "Xuefeng Xu"},
]
description = "Federated data Preprocessing with Sketching"
readme = "README.md"
license = {file = "LICENSE"}
keywords = [
    "Data Preprocessing",
    "Federated Learning",
    "Data Sketching",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
]

[project.urls]
Source = "https://github.com/primihub/fedps"
Tracker = "https://github.com/primihub/fedps/issues"
